+++ src/hifivar/__init__.py
@@ -1,5 +1,5 @@
 """HiFiVar package."""
 
-__version__ = "0.1.0rc3"
+__version__ = "0.1.0rc4"
 
 __all__ = ["__version__"]